refactor(AdminReservationController): endpoint를 고려하여 컨트롤러 분리
