+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,7 +1,6 @@
 package roomescape.controller;
 
 import java.net.URI;
-import java.time.LocalDate;
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
@@ -10,11 +9,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import roomescape.annotation.AuthenticationPrincipal;
-import roomescape.controller.request.AdminReservationRequest;
 import roomescape.controller.request.ReservationRequest;
 import roomescape.controller.response.ReservationResponse;
 import roomescape.model.Reservation;
@@ -37,24 +34,6 @@ public class ReservationController {
                 .map(ReservationResponse::new)
                 .toList();
         return ResponseEntity.ok(responses);
-    }
-
-    @GetMapping("/admin/reservations")
-    public ResponseEntity<List<ReservationResponse>> searchReservations(@RequestParam("themeId") Long themeId,
-                                                                        @RequestParam("memberId") Long memberId,
-                                                                        @RequestParam("dateFrom") LocalDate dateFrom,
-                                                                        @RequestParam("dateTo") LocalDate dateTo) {
-        List<Reservation> reservations = reservationService.filterReservation(themeId, memberId, dateFrom, dateTo);
-        List<ReservationResponse> responses = reservations.stream()
-                .map(ReservationResponse::new)
-                .toList();
-        return ResponseEntity.ok(responses);
-    }
-
-    @PostMapping("/admin/reservations")
-    public ResponseEntity<Reservation> createReservation(@RequestBody AdminReservationRequest request) {
-        Reservation reservation = reservationService.addReservation(request);
-        return ResponseEntity.created(URI.create("/reservations/" + reservation.getId())).body(reservation);
     }
 
     @PostMapping("/reservations")
